@@ -1,4 +1,5 @@
 import 'dart:async';
+
 import 'package:flutter/services.dart';
 
 const EventChannel _accelerometerEventChannel =
@@ -7,11 +8,9 @@ const EventChannel _accelerometerEventChannel =
 const EventChannel _userAccelerometerEventChannel =
     EventChannel('cindyu.com/all_sensors/user_accel');
 
-const EventChannel _gyroscopeEventChannel =
-    EventChannel('cindyu.com/all_sensors/gyroscope');
+const EventChannel _gyroscopeEventChannel = EventChannel('cindyu.com/all_sensors/gyroscope');
 
-const EventChannel _proximityEventChannel =
-    EventChannel('cindyu.com/all_sensors/proximity');
+const EventChannel _proximityEventChannel = EventChannel('cindyu.com/all_sensors/proximity');
 
 class AccelerometerEvent {
   /// Acceleration force along the x axis (including gravity) measured in m/s^2.
@@ -93,24 +92,23 @@ ProximityEvent _listToProximityEvent(List<double> list) {
   return new ProximityEvent(list[0]);
 }
 
-Stream<AccelerometerEvent> _accelerometerEvents;
-Stream<GyroscopeEvent> _gyroscopeEvents;
-Stream<UserAccelerometerEvent> _userAccelerometerEvents;
-Stream<ProximityEvent> _proximityEvents;
+Stream<AccelerometerEvent>? _accelerometerEvents;
+Stream<GyroscopeEvent>? _gyroscopeEvents;
+Stream<UserAccelerometerEvent>? _userAccelerometerEvents;
+Stream<ProximityEvent>? _proximityEvents;
 
 /// A broadcast stream of events from the device accelerometer.
-Stream<AccelerometerEvent> get accelerometerEvents {
+Stream<AccelerometerEvent>? get accelerometerEvents {
   if (_accelerometerEvents == null) {
     _accelerometerEvents = _accelerometerEventChannel
         .receiveBroadcastStream()
-        .map(
-            (dynamic event) => _listToAccelerometerEvent(event.cast<double>()));
+        .map((dynamic event) => _listToAccelerometerEvent(event.cast<double>()));
   }
   return _accelerometerEvents;
 }
 
 /// A broadcast stream of events from the device gyroscope.
-Stream<GyroscopeEvent> get gyroscopeEvents {
+Stream<GyroscopeEvent>? get gyroscopeEvents {
   if (_gyroscopeEvents == null) {
     _gyroscopeEvents = _gyroscopeEventChannel
         .receiveBroadcastStream()
@@ -120,19 +118,17 @@ Stream<GyroscopeEvent> get gyroscopeEvents {
 }
 
 /// Events from the device accelerometer with gravity removed.
-Stream<UserAccelerometerEvent> get userAccelerometerEvents {
+Stream<UserAccelerometerEvent>? get userAccelerometerEvents {
   if (_userAccelerometerEvents == null) {
     _userAccelerometerEvents = _userAccelerometerEventChannel
         .receiveBroadcastStream()
-        .map((dynamic event) =>
-            _listToUserAccelerometerEvent(event.cast<double>()));
+        .map((dynamic event) => _listToUserAccelerometerEvent(event.cast<double>()));
   }
   return _userAccelerometerEvents;
 }
 
 /// A broadcast stream of events from the device proximity.
-Stream<ProximityEvent> get proximityEvents {
-
+Stream<ProximityEvent>? get proximityEvents {
   _proximityEvents = _proximityEventChannel
       .receiveBroadcastStream()
       .map((dynamic event) => _listToProximityEvent(event.cast<double>()));
